Updates 1. abstracted home page into modular components 2. abstracted login page into reusable components 3. abstracted registration page into reusable components
