+++ expense-tracker-frontend/src/components/common/theme/darkTheme.js
@@ -1,0 +1,16 @@
+import { createTheme } from '@mui/material/styles';
+
+const darkTheme = createTheme({
+  palette: {
+    mode: 'dark',
+    primary: {
+      main: '#2563eb',
+    },
+    background: {
+      default: '#030712',
+      paper: '#111827',
+    },
+  },
+});
+
+export default darkTheme;
